sbxg: introduce menuconfig and uniform config
Configuration can be done thanks to conf/mconf (menuconfig).
menuconfig will fetch in the manifests directory which boards
and manifests are available and present them to the user.

Configuration files are now .config. Exit config.template and
config.user. Now, generate a configuration with menuconfig or
defconfig, and eventually edit manually the generated .config.
This file may be store and re-used later.