make: fix clean targets
So, we got this:

   $ make V=1 clean
   make -C u-boot CROSS_COMPILE="arm-linux-gnueabihf-" -j 3 .o
   make[1]: Entering directory '/local_home/local/git/SBXG/sbxg.git/u-boot'
   o .o
   make[1]: o: commande introuvable
   <builtin>: recipe for target '.o' failed
   make[1]: [.o] Error 127 (ignorée)
   make[1]: Leaving directory '/local_home/local/git/SBXG/sbxg.git/u-boot'
   cc   u-boot-.o   -o u-boot-
   gcc-4.9.real: error: u-boot-.o: Aucun fichier ou dossier de ce type
   gcc-4.9.real: fatal error: no input files
   compilation terminated.
   <builtin>: recipe for target 'u-boot-' failed
   make: *** [u-boot-] Error 4

It appears using $@ in a dependency name may result in an undefined
behaviour. I have no reference to prove that, but replacing $@ by the
expected target name before this patch was sufficient to fix this.
So, by moving the sub-clean targets in a command, it has the expected
behaviour.