sbxg: add the local toolchain
The local toolchain introduced a paradigm shift in SBXG. Components were
linked to a toolchain, but this design was plain wrong. Linking a board
to a toolchain implied it could only be built for the host architecture
of the cross-compiler!

Now, boards depend on an architecture that the toolchain (build) must
fulfill. This includes the local toolchain, that does not require
fetching.